% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{ConstantTriangle}
\alias{ConstantTriangle}
\title{Constant Triangle}
\format{A list of five 3x3 matrices.}
\description{
A list of distance matrices representing a constant equilateral triangle at five timesteps.
}
\keyword{datasets}

