\docType{data}
\name{ExpandingCircle}
\alias{ExpandingCircle}
\title{Expanding Circle}
\format{A list of ten 10x10 matrices.}
\description{
  A list of distance matrices representing ten points
  arranged on a circle with equal spacing. The circle has
  linearly increasing radius. There is a bit gaussian noise
  on the position of the points.
}
\keyword{datasets}

