\docType{data}
\name{ExpandingTriangle}
\alias{ExpandingTriangle}
\title{Expanding Triangle}
\format{A list of five 3x3 matrices.}
\description{
  A list of distance matrices representing three points on
  the corner of an equilateral triangle. The side lenghts
  of the triangle are linearly increasing.
}
\keyword{datasets}

